requirements.txt
gym==0.26.2
numpy==1.23.5
pandas==1.5.3
torch==1.13.1
argparse
PyYAML>=5.3.1
scikit-learn>=1.0.0
yfinance>=0.2.0
ta>=0.8.0
seaborn>=0.11.0
matplotlib>=3.4.0
pytest>=6.0.0
pyarrow>=8.0.0
optuna>=3.0.0
ccxt>=3.0.0
python-dateutil>=2.8.0
